Title: 사각형 채우기 3, Time: 105ms, Memory: 9MB, Status: Passed - CodeTree

+++ 240805/사각형 채우기 3/rectangle-fill-3.java
@@ -7,7 +7,7 @@ public class Main {
     static StringBuilder sb = new StringBuilder();
 
     static int n;
-    static int[] dp = new int[1001];
+    static long[] dp = new long[1001];
 
     public static void main(String[] args) throws Exception {
